requirements.txt
Flask==2.2.2
werkzeug==2.2.2
psycopg2-binary==2.9.3
Faker==18.2
